linux: Detect the battery of bluetooth input devices
Check the input subdevices of the bluetooth parent device of a battery, if
present. If there is mouse* input child device, is a mouse battery; otherwise,
it is a keyboard battery. This also fixes the PowerSupply attribute for these
to be false, as the batteries of wireless input devices don't power the system.

https://launchpad.net/bugs/1153488

Signed-Off-By: Martin Pitt <martin.pitt@ubuntu.com>