UpKbdBacklight: don't cache the brightness level, always read it from sysfs attrib
When GetBrightness is called it's better to fetch the current value in the
sysfs attribute since many devices doesn't support the emission of brightness
changes to userland, and this could make settings daemons to handle
this value incorrectly.

Add static up_kbd_backlight_brightness_read that is now called during
initialization too, it returns -1 on errors, and use it everywhere we need
to read or check the current brightness.

https://bugs.freedesktop.org/show_bug.cgi?id=95457