etc: Tweak default percentage levels
If we want the computer to be able to take useful action about the low
battery, we should have a slightly higher "low" percentage level so that
power saving made really makes a difference in runtime.

Also bump "critical" slightly so that doom isn't quite as near but in the
distance nonetheless.

The "action" level stays the same, as 1% is too close to some batteries'
actual switch off point, eg. the computer might brown out before we see
1%.