rules: Update udev rules and migrate to use hwdb
This makes rule matching more efficient. We just need to copy the
variables over into the hiddev device so that there is no need to change
the C code to match the parent USB device. That said, the change in C
wouldn't be that hard either.

See https://github.com/networkupstools/nut/pull/1342