daemon: Fix reference counting at daemon shutdown
Commit b78a2ab980ca ("daemon: Stop passing UpDevice reference to the
daemon") fixed the fact that the signal passed a reference to the
daemon. This fix also implied that the UpDaemon object does not own a
reference to the device itself, instead relying on the UpDeviceList to
keep it alive.

With this change, there is no need to drop the extra reference when
clearing the device list, remove the code accordingly.

Reported-by: Alberts Muktupāvels <alberts.muktupavels@gmail.com>