daemon: Update aggregate state matching and handle UNKNOWN state
Stop trying to guess a battery state based on the AC information for a
single battery. Instead, just do the guessing in the display device.

This means that cases with more than one battery work fine. It also
means that we still report an UNKNOWN state for the battery itself.

Also get the on_battery information from the display devices discharging
state, there is no need to iterate all of the batteries to figure that
out.

Note that we don't really have a well defined state for the display
device in all cases. We'll add a test in the next step, marking those
cases as TBD or ANY to show where the issues are. AFAICT we do not have
a regression though, so this should be fine for now.

Fixes: #146