battery: User power/current reading from battery
This switches to always use the power/current reading if we had a
sensible reading at some point in the past.

In contrast to the older UPower code, will will however ignore the
reading for 10 seconds after a discontinuity (power plug/unplug or
resume) unless we read an explicit zero value after the event happened.

We do this under the assumption that the readings may be wildly off, and
it is better to not show an estimate rather than one that is wildly
incorrect.

Note that this commit also normalises negative power readings while
discharging to be positive.

Closes: #199