linux: Fix warning when a duplicate device appears
When a kernel device appears after the Bluetooth device, we could be
trying to bind the model property onto itself on the Bluetooth device.
Avoid this possibility by always binding the property from the bluez
device to the non-bluez device.

GLib-GObject-WARNING **: 20:20:50.644: Unable to bind the same property on the same instance

Fixes: 778b93a336bef ("linux: Hide duplicate Logitech Bluetooth devices")