linux: Hide duplicate USB devices
Make our unduplicating code a bit more clever and hide the non-BlueZ
devices with the "unknown" battery state.

When Logitech devices are connected through Unifying and USB, with the
correct Linux patches[1], their serial numbers match, but the Unifying
connection stops exporting the battery information, so hide that one
from the export.

Closes: #206

[1]: https://patchwork.kernel.org/project/linux-input/list/?series=726116