up-daemon: Handle the AC is online but the battery is discharging      
If a low-power charger connects to the system, the battery still reports
discharging since the AC charger can't provide enough power to make      
the system runs and charges the battery at the same time. In this situation,
we estimate the warning level to trigger the CriticalPowerAction until
the battery status is changed to charging.

Resolves: #276

Signed-off-by: Kate Hsuan <hpa@redhat.com>