up-daemon: determining a low-efficiency charger through monitoring the battery percentage
Some of the battery controller discharges the battery when the percentage
reaches the charge_control_end_threshold. It discharges the battery to a certain
percentage and then starts to charge the battery to the end threshold. Therefore,
the charging status is frequently changed.

If the battery percentage drops 3% in a period (300 secs), upower sets the
on_battery to yes to reflect the real battery charging status and upower
assumes a low-efficiency charger is attached. If not, the counter will be
reset.