Revert "up-daemon: Handle the AC is online but the battery is discharging"
The commit being reverted introduces the bug #306 by claiming external
power is not available when the device battery begins to discharge for
any reason.

We fix this issue by reverting to the old behaviour of reporting
external power being available as long as it's online. We then fix
issue #276 via alternative means.

Resolves: #306

This reverts commit 9d754d8b2fbc644e7a865d828bffcd58e25c49d2.